// frontend/src/types.ts
export interface CustomFileBase extends Partial<globalThis.File> {
  processingTotalTime: number | string;
  status: string;
  nodesCount: number;
  relationshipsCount: number;
  model: string;
  fileSource: string;
  sourceUrl?: string;
  wikiQuery?: string;
  gcsBucket?: string;
  gcsBucketFolder?: string;
  errorMessage?: string;
  uploadProgress?: number;
  processingStatus?: boolean;
  googleProjectId?: string;
  language?: string;
  processingProgress?: number;
  accessToken?: string;
  isChecked?: boolean;
  retryOptionStatus: boolean;
  retryOption: string;
  chunkNodeCount: number;
  chunkRelCount: number;
  entityNodeCount: number;
  entityEntityRelCount: number;
  communityNodeCount: number;
  communityRelCount: number;
  createdAt?: Date;
  token_usage: number;
  embedding_model?: string;
}
export interface CustomFile extends CustomFileBase {
  id: string;
}
// ...
export interface ScanProps {
  urlParam?: string;
  model?: string;
  accessKey?: string;
  secretKey?: string;
  wikiquery?: string;
  gcs_bucket_name?: string;
  gcs_bucket_folder?: string;
  source_type?: string;
  gcs_project_id?: string;
  access_token?: string;
}